+++ auto_route/example/lib/mobile/router/router.dart
@@ -1,8 +1,10 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:example/mobile/screens/books/book_details_page.dart';
 import 'package:example/mobile/screens/books/book_list_page.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/foundation.dart' show kIsWeb;
+import 'package:flutter/material.dart';
 
-import '../screens/books/routes.dart';
 import '../screens/home_page.dart';
 import '../screens/login_page.dart';
 import '../screens/profile/routes.dart';
@@ -12,10 +14,10 @@ import '../screens/user-data/routes.dart';
 export 'router.gr.dart';
 
 @MaterialAutoRouter(
-  replaceInRouteName: 'Page,Route',
+  replaceInRouteName: 'Page|Dialog,Route',
   routes: <AutoRoute>[
     // app stack
-    AutoRoute(
+    AutoRoute<String>(
       path: '/',
       page: HomePage,
       children: [
@@ -25,12 +27,7 @@ export 'router.gr.dart';
           name: 'BooksTab',
           children: [
             AutoRoute(path: '', page: BookListPage),
-            AutoRoute(
-              path: ':id',
-              usesPathAsKey: true,
-              page: BookDetailsPage,
-              // guards: [AuthGuard],
-            ),
+            AutoRoute(path: ':id', page: BookDetailsPage),
           ],
         ),
         profileTab,
@@ -43,11 +40,52 @@ export 'router.gr.dart';
     ),
     userDataRoutes,
     // auth
-    AutoRoute(
-      path: '/login',
-      page: LoginPage,
-    ),
+
+    AutoRoute(page: LoginPage, path: '/login'),
+
     RedirectRoute(path: '*', redirectTo: '/'),
   ],
 )
 class $RootRouter {}
+
+// CustomRoute is coming from auto_route
+class DialogModalRoute<T> extends CustomRoute<T> {
+  const DialogModalRoute({required Type page, String? path})
+      : super(
+          page: page,
+          path: path,
+          // customRouteBuilder: dialogRouteBuilder,
+        );
+
+  // must be static and public
+  // static Route<T> dialogRouteBuilder<T>(
+  //   BuildContext context,
+  //   Widget child,
+  //   // CustomPage<T> page,
+  // ) {
+  //   // DialogRoute is coming from flutter material
+  //   if (kIsWeb) {
+  //     return DialogRoute<T>(
+  //       context: context,
+  //       settings: page,
+  //       builder: (context) => child,
+  //     );
+  //   } else {
+  //     return MaterialPageRoute<T>(
+  //       settings: page,
+  //       builder: (context) => child,
+  //     );
+  //   }
+  // }
+}
+
+// typedef CustomRouteBuilder = Route<T> Function<T>(BuildContext context, Widget child, CustomPage page);
+//
+// Route<T> myCustomRouteBuilder<T>(
+//     BuildContext context, Widget child, CustomPage<T> page) {
+//   return PageRouteBuilder(
+//       fullscreenDialog: page.fullscreenDialog,
+//       // this's important
+//       settings: page,
+//       pageBuilder: (_, __, ___) => child);
+// }