+++ auto_route/example/lib/mobile/screens/profile/my_books_page.dart
@@ -1,19 +1,16 @@
 import 'package:auto_route/auto_route.dart';
+import 'package:example/mobile/router/router.gr.dart';
 import 'package:flutter/material.dart';
 
 class MyBooksPage extends StatelessWidget {
-  final String filter;
+  final String? filter;
 
-  MyBooksPage({Key? key, @queryParam this.filter = 'none'}) : super(key: key) {
-    print('constructing myBooksPage');
-  }
+  MyBooksPage({Key? key, @queryParam this.filter = 'none'}) : super(key: key) {}
 
   @override
   Widget build(BuildContext context) {
-    print('building myBooksPage');
-    print(context.routeData.queryParams);
-    return Container(
-      child: Center(
+    return Scaffold(
+      body: Center(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
@@ -29,10 +26,9 @@ class MyBooksPage extends StatelessWidget {
             const SizedBox(height: 32),
             ElevatedButton(
               onPressed: () {
-                context.router.navigateNamed('/settings/newSegment');
-                //   return context.router.navigate(
-                //   BooksTab(),
-                // );
+                context.navigateTo(
+                  SettingsTab(tab: 'newSegment'),
+                );
               },
               child: Text('navigate to /settings/newSegment'),
             )