+++ auto_route/example/lib/mobile/screens/profile/routes.dart
@@ -1,7 +1,7 @@
 import 'package:auto_route/annotations.dart';
 import 'package:auto_route/auto_route.dart';
+import 'package:example/mobile/screens/profile/my_books_page.dart';
 
-import 'my_books_page.dart';
 import 'profile_page.dart';
 
 const profileTab = AutoRoute(