+++ auto_route/example/lib/mobile/screens/user-data/routes.dart
@@ -1,29 +1,29 @@
 import 'package:auto_route/auto_route.dart';
-import 'package:example/mobile/screens/user-data/sinlge_field_page.dart';
+import 'package:example/mobile/screens/user-data/single_field_page.dart';
 import 'package:example/mobile/screens/user-data/user_data_page.dart';
 
 import 'data_collector.dart';
 
 const userDataRoutes = AutoRoute<UserData>(
-  path: 'user-data',
+  path: '/user-data',
   page: UserDataCollectorPage,
   children: [
-    CustomRoute(
+    AutoRoute(
       path: 'name',
       name: 'NameFieldRoute',
       page: SingleFieldPage,
-      transitionsBuilder: TransitionsBuilders.slideRightWithFade,
+      // transitionsBuilder: TransitionsBuilders.slideRightWithFade,
     ),
-    CustomRoute(
+    AutoRoute(
       path: 'favorite-book',
       page: SingleFieldPage,
       name: 'FavoriteBookFieldRoute',
-      transitionsBuilder: TransitionsBuilders.slideRightWithFade,
+      // transitionsBuilder: TransitionsBuilders.slideRightWithFade,
     ),
-    CustomRoute(
+    AutoRoute(
       path: 'results',
       page: UserDataPage,
-      transitionsBuilder: TransitionsBuilders.slideRightWithFade,
+      // transitionsBuilder: TransitionsBuilders.slideRightWithFade,
     ),
   ],
 );