+++ auto_route/lib/src/router/widgets/auto_back_button.dart
@@ -3,22 +3,23 @@ import 'package:flutter/material.dart';
 
 import '../../../auto_route.dart';
 
-class AutoBackButton extends StatefulWidget {
+class AutoBackButton extends StatelessWidget {
   final Color? color;
-  const AutoBackButton({Key? key, this.color}) : super(key: key);
+  final bool showIfParentCanPop;
+  const AutoBackButton({
+    Key? key,
+    this.color,
+    this.showIfParentCanPop = true,
+  }) : super(key: key);
 
-  @override
-  _AutoBackButtonState createState() => _AutoBackButtonState();
-}
-
-class _AutoBackButtonState extends State<AutoBackButton> {
   @override
   Widget build(BuildContext context) {
-    final scope = RouterScope.of(context);
-    if (scope.controller.canPopSelfOrChildren) {
+    final scope = AutoRouter.of(context, watch: true);
+    if (scope.canPopSelfOrChildren ||
+        (showIfParentCanPop && scope.parent()?.canPopSelfOrChildren == true)) {
       return BackButton(
-        color: widget.color,
-        onPressed: scope.controller.popTop,
+        color: color,
+        onPressed: scope.popTop,
       );
     }
     return const SizedBox.shrink();