+++ auto_route/lib/src/router/widgets/auto_route_navigator.dart
@@ -1,4 +1,5 @@
 import 'package:collection/collection.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import '../../../auto_route.dart';
@@ -32,9 +33,19 @@ class _AutoRouteNavigatorState extends State<AutoRouteNavigator> {
   void didChangeDependencies() {
     super.didChangeDependencies();
     if (widget.declarativeRoutesBuilder != null && _routesSnapshot == null) {
-      var routes = widget.declarativeRoutesBuilder!(context);
-      _routesSnapshot = routes;
-      widget.router.updateDeclarativeRoutes(routes);
+      _updateDeclarativeRoutes();
+    }
+  }
+
+  void _updateDeclarativeRoutes() {
+    final delegate = AutoRouterDelegate.of(context);
+    var newRoutes = widget.declarativeRoutesBuilder!(context);
+    if (!ListEquality().equals(newRoutes, _routesSnapshot)) {
+      _routesSnapshot = newRoutes;
+      widget.router.updateDeclarativeRoutes(newRoutes);
+      WidgetsBinding.instance?.addPostFrameCallback((_) {
+        delegate.notifyUrlChanged();
+      });
     }
   }
 
@@ -42,34 +53,20 @@ class _AutoRouteNavigatorState extends State<AutoRouteNavigator> {
   void didUpdateWidget(covariant AutoRouteNavigator oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (widget.declarativeRoutesBuilder != null) {
-      var shouldNotify = false;
-      final delegate = AutoRouterDelegate.of(context);
-      var newRoutes = widget.declarativeRoutesBuilder!(context);
-      if (!ListEquality().equals(newRoutes, _routesSnapshot)) {
-        shouldNotify = true;
-        _routesSnapshot = newRoutes;
-        widget.router.updateDeclarativeRoutes(newRoutes);
-      } else if (!ListEquality().equals(
-        delegate.urlState.segments,
-        delegate.controller.currentSegments,
-      )) {
-        shouldNotify = true;
-      }
-      if (shouldNotify) {
-        WidgetsBinding.instance?.addPostFrameCallback((_) {
-          delegate.notifyUrlChanged();
-        });
-      }
+      _updateDeclarativeRoutes();
     }
   }
 
   @override
   Widget build(BuildContext context) {
-    return Navigator(
+    final navigator = Navigator(
       key: widget.router.navigatorKey,
       observers: widget.navigatorObservers,
-      restorationScopeId: widget.navRestorationScopeId ?? widget.router.routeData.name,
-      pages: widget.router.hasEntries ? widget.router.stack : [_PlaceHolderPage(widget.placeholder)],
+      restorationScopeId:
+          widget.navRestorationScopeId ?? widget.router.routeData.name,
+      pages: widget.router.hasEntries
+          ? widget.router.stack
+          : [_PlaceHolderPage(widget.placeholder)],
       transitionDelegate: _CustomTransitionDelegate(),
       onPopPage: (route, result) {
         if (!route.didPop(result)) {
@@ -83,6 +80,18 @@ class _AutoRouteNavigatorState extends State<AutoRouteNavigator> {
         return true;
       },
     );
+
+    // fixes nested cupertino routes back gesture issue
+    if (!widget.router.isRoot) {
+      return WillPopScope(
+        child: navigator,
+        onWillPop: widget.router.canPopSelfOrChildren
+            ? () => SynchronousFuture(true)
+            : null,
+      );
+    }
+
+    return navigator;
   }
 }
 
@@ -116,35 +125,44 @@ class _CustomTransitionDelegate<T> extends TransitionDelegate<T> {
   @override
   Iterable<RouteTransitionRecord> resolve({
     required List<RouteTransitionRecord> newPageRouteHistory,
-    required Map<RouteTransitionRecord?, RouteTransitionRecord> locationToExitingPageRoute,
-    required Map<RouteTransitionRecord?, List<RouteTransitionRecord>> pageRouteToPagelessRoutes,
+    required Map<RouteTransitionRecord?, RouteTransitionRecord>
+        locationToExitingPageRoute,
+    required Map<RouteTransitionRecord?, List<RouteTransitionRecord>>
+        pageRouteToPagelessRoutes,
   }) {
     final List<RouteTransitionRecord> results = <RouteTransitionRecord>[];
     // This method will handle the exiting route and its corresponding pageless
     // route at this location. It will also recursively check if there is any
     // other exiting routes above it and handle them accordingly.
     void handleExitingRoute(RouteTransitionRecord? location, bool isLast) {
-      final RouteTransitionRecord? exitingPageRoute = locationToExitingPageRoute[location];
+      final RouteTransitionRecord? exitingPageRoute =
+          locationToExitingPageRoute[location];
       if (exitingPageRoute == null) return;
       if (exitingPageRoute.isWaitingForExitingDecision) {
-        final bool hasPagelessRoute = pageRouteToPagelessRoutes.containsKey(exitingPageRoute);
-        final bool isLastExitingPageRoute = isLast && !locationToExitingPageRoute.containsKey(exitingPageRoute);
+        final bool hasPagelessRoute =
+            pageRouteToPagelessRoutes.containsKey(exitingPageRoute);
+        final bool isLastExitingPageRoute =
+            isLast && !locationToExitingPageRoute.containsKey(exitingPageRoute);
         if (isLastExitingPageRoute && !hasPagelessRoute) {
           exitingPageRoute.markForPop(exitingPageRoute.route.currentResult);
         } else {
-          exitingPageRoute.markForComplete(exitingPageRoute.route.currentResult);
+          exitingPageRoute
+              .markForComplete(exitingPageRoute.route.currentResult);
         }
         if (hasPagelessRoute) {
-          final List<RouteTransitionRecord> pagelessRoutes = pageRouteToPagelessRoutes[exitingPageRoute]!;
+          final List<RouteTransitionRecord> pagelessRoutes =
+              pageRouteToPagelessRoutes[exitingPageRoute]!;
           for (final RouteTransitionRecord pagelessRoute in pagelessRoutes) {
             // It is possible that a pageless route that belongs to an exiting
             // page-based route does not require exiting decision. This can
             // happen if the page list is updated right after a Navigator.pop.
             if (pagelessRoute.isWaitingForExitingDecision) {
-              if (isLastExitingPageRoute && pagelessRoute == pagelessRoutes.last) {
+              if (isLastExitingPageRoute &&
+                  pagelessRoute == pagelessRoutes.last) {
                 pagelessRoute.markForPop(pagelessRoute.route.currentResult);
               } else {
-                pagelessRoute.markForComplete(pagelessRoute.route.currentResult);
+                pagelessRoute
+                    .markForComplete(pagelessRoute.route.currentResult);
               }
             }
           }
@@ -161,9 +179,12 @@ class _CustomTransitionDelegate<T> extends TransitionDelegate<T> {
 
     for (final RouteTransitionRecord pageRoute in newPageRouteHistory) {
       final bool isLastIteration = newPageRouteHistory.last == pageRoute;
-      final bool firstPageIsPlaceHolder = results.isNotEmpty && results.first.route.settings is _PlaceHolderPage;
+      final bool firstPageIsPlaceHolder = results.isNotEmpty &&
+          results.first.route.settings is _PlaceHolderPage;
       if (pageRoute.isWaitingForEnteringDecision) {
-        if (!locationToExitingPageRoute.containsKey(pageRoute) && isLastIteration && !firstPageIsPlaceHolder) {
+        if (!locationToExitingPageRoute.containsKey(pageRoute) &&
+            isLastIteration &&
+            !firstPageIsPlaceHolder) {
           pageRoute.markForPush();
         } else {
           pageRoute.markForAdd();