+++ auto_route/lib/src/router/widgets/auto_router_delegate.dart
@@ -3,6 +3,7 @@ import 'package:auto_route/src/matcher/route_matcher.dart';
 import 'package:auto_route/src/route/page_route_info.dart';
 import 'package:auto_route/src/router/controller/controller_scope.dart';
 import 'package:auto_route/src/router/parser/route_information_parser.dart';
+import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
@@ -48,7 +49,9 @@ class AutoRouterDelegate extends RouterDelegate<UrlState> with ChangeNotifier {
   }
 
   @override
-  Future<bool> popRoute() => controller.topMost.pop();
+  Future<bool> popRoute() {
+    return controller.topMost.pop();
+  }
 
   late List<NavigatorObserver> _navigatorObservers;
 
@@ -78,14 +81,7 @@ class AutoRouterDelegate extends RouterDelegate<UrlState> with ChangeNotifier {
   UrlState get urlState => _urlState;
 
   @override
-  UrlState? get currentConfiguration {
-    final newState = UrlState.fromSegments(controller.currentSegments);
-    if (_urlState != newState) {
-      _urlState = newState;
-      return newState;
-    }
-    return null;
-  }
+  UrlState? get currentConfiguration => _urlState;
 
   @override
   Future<void> setInitialRoutePath(UrlState tree) {
@@ -111,9 +107,9 @@ class AutoRouterDelegate extends RouterDelegate<UrlState> with ChangeNotifier {
   }
 
   @override
-  Future<void> setNewRoutePath(UrlState tree) {
-    if (tree.hasSegments) {
-      return controller.navigateAll(tree.segments);
+  Future<void> setNewRoutePath(UrlState state) {
+    if (state.hasSegments) {
+      return controller.navigateAll(state.segments);
     }
     return SynchronousFuture(null);
   }
@@ -140,7 +136,17 @@ class AutoRouterDelegate extends RouterDelegate<UrlState> with ChangeNotifier {
   }
 
   void _rebuildListener() {
-    notifyListeners();
+    final newState = UrlState.fromSegments(controller.currentSegments);
+    if (_urlState.url != newState.url) {
+      final segments = newState.segments;
+
+      final replace = segments.isNotEmpty &&
+          (segments.last.fromRedirect ||
+              (segments.last.hasEmptyPath && _urlState.path == '/'));
+
+      _urlState = newState.copyWith(replace: replace);
+      notifyListeners();
+    }
   }
 
   @override
@@ -150,7 +156,8 @@ class AutoRouterDelegate extends RouterDelegate<UrlState> with ChangeNotifier {
   }
 
   void notifyUrlChanged() {
-    notifyListeners();
+    // notifyListeners();
+    _rebuildListener();
   }
 }